rename spm_config.py to avoid confusion
