+++ bids_prov/tests/test_spm_parser.py
@@ -1,7 +1,7 @@
 import pytest
 import json
 
-from ..spm_config import DEPENDENCY_REGEX
+from ..spm_load_config import DEPENDENCY_REGEX
 from ..spm_parser import get_records, get_closest_activity
 import re
